[QA] scmport.h, scmport-config.h, scmport-null.h, scmport-null.c,      scmport-file.h, scmport-file.c, scmport-str.h, scmport-str.c,      scmport-basechar.c, scmport-sbchar.h, scmport-sbchar.c
* sigscheme/src/scmport.h
  - (SCM_PORT_ASSERT): New macro
  - (SCM_CHARPORT_ENCODING, SCM_CHARPORT_CCS): Support
    !SCM_USE_MULTIBYTE_CHAR without encoding.[hc]
  - (ScmCharCodec, SCM_CCS_UNKNOWN): New macro for
    !SCM_USE_MULTIBYTE_CHAR
  - (struct ScmBaseCharPort_): Change type of linenum from int to
    size_t, to avoid potencial overflow and adjust alignment
  - (ScmBaseCharPort_line_number): Change return type to size_t
  - Cosmetic changes
* sigscheme/src/scmport-sbchar.c
  - (struct ScmSingleByteCharPort_): Fix broken ScmBaseCharPort
    inheritance which may cause SEGV
  - (scm_sbcport_init): Support !SCM_USE_MULTIBYTE_CHAR without encoding.[hc]
  - (sbcport_put_char): Make logically correct
* sigscheme/src/scmport-basechar.c
  - (ScmBaseCharPort_line_number): Change return type to size_t
  - (ScmBaseCharPort_construct): Change invalid linenum from -1 to 0
* sigscheme/src/scmport-mbchar.c
  - (struct ScmMultiByteCharPort_): Follow the type change of linenum
* sigscheme/src/read.c
  - (read_list):
    * Follow the type change of ScmBaseCharPort_line_number()
    * Make format needed only if SCM_DEBUG
* sigscheme/src/scmport-file.c
  - (fileport_new_internal): Add some assertions as precondition
* sigscheme/src/scmport-str.c
  - (istrport_new, ostrport_append): Add an assertions as precondition
* sigscheme/src/scmport-config.h
  - (SCM_PORT_CALLOC, SCM_PORT_REALLOC): Fix macro definition style
* sigscheme/src/sigscheme.h
  - (SCM_ASSERT): Use assert(3) only if HAVE_ASSERT_H &&
    !SCM_CHICKEN_DEBUG
* sigscheme/configure.in
  - Add detection for assert.h
* sigscheme/QALog
  - Update

+++ sigscheme/src/scmport-mbchar.c
@@ -68,7 +68,7 @@ struct ScmMultiByteCharPort_ {  /* inherits ScmBaseCharPort */
     const ScmCharPortVTbl *vptr;
 
     ScmBytePort *bport;  /* protected */
-    int linenum;         /* protected */
+    size_t linenum;      /* protected */
 
     ScmCharCodec *codec;
     ScmMultibyteState state;
